requirements.txt
-r lyrics-generator_copy/requirements_base.txt

absl-py==2.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.2.0
beautifulsoup4==4.12.2
better-profanity==0.6.1
billiard==4.2.0
blinker==1.7.0
boto==2.49.0
bs4==0.0.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==42.0.2
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.44.0
dvc-data==3.11.0
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.0.0
dvc-render==1.0.1
dvc-studio-client==0.19.0
dvc-task==0.3.0
entrypoints==0.4
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==2.0.7
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2024.2.0
funcy==2.0
gast==0.4.0
gensim==3.8.0
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-compute-engine==2.8.3
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grandalf==0.8
grpcio==1.60.1
gto==1.7.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==7.0.1
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.9.4
joblib==1.3.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kombu==5.3.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
networkx==3.2.1
numpy==1.25.2
oauth2client==4.1.3
omegaconf==2.3.0
opt-einsum==3.3.0
orjson==3.9.13
packaging==23.2
pandas==1.1.5
pathspec==0.12.1
platformdirs==3.11.0
prompt-toolkit==3.0.43
protobuf>3.20.1, <=3.21.0
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydot==2.0.0
PyDrive2==1.19.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==0.24.2
scipy==1.5.4
scmrepo==3.1.0
semver==3.0.2
shortuuid==1.0.11
shtab==1.6.5
six==1.15.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
sqltrie==0.11.0
tabulate==0.9.0
telegram_text==0.2.0
tensorboard==2.16.0
tensorboard-data-server==0.7.2
tensorflow>=2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.9.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-text==2.7.3
tensorflowjs==3.6.0
termcolor==2.4.0
tf_keras-nightly==2.16.0.dev2024021210
threadpoolctl==3.2.0
tomlkit==0.12.3
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.5
validators==0.22.0
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
